Remove more uses of %n. (dm)
git-svn-id: svn://mielke.cc/main/brltty@6458 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/auth.h
@@ -29,7 +29,7 @@ extern AuthDescriptor *authBeginServer (const char *parameter);
 extern void authEnd (AuthDescriptor *auth);
 extern int authPerform (AuthDescriptor *auth, FileDescriptor fd);
 
-extern void formatAddress (char *buffer, int bufferSize, const void *address, int addressSize);
+extern void formatAddress (char *buffer, size_t bufferSize, const void *address, int addressSize);
 
 #ifdef __cplusplus
 }
